+++ mm485/arduino/codec128.h
@@ -10,8 +10,8 @@
 
 #include "Arduino.h"
 
-int enc128(unsigned char* buf, const unsigned char* data, size_t size);
-int dec128(unsigned char* buf, const unsigned char* data, size_t size);
+uint8_t enc128(unsigned char* buf, const unsigned char* data, uint8_t size);
+uint8_t dec128(unsigned char* buf, const unsigned char* data, uint8_t size);
 
 
 #endif /* MM485_ARDUINO_CODEC128_H_ */
